requirements.txt
amberelectric==2.0.12
dataclasses_json
flask
pytest
python-dateutil
python-dotenv==1.1.1
requests==2.32.4
supervisor
